Minor housekeeping

+++ special_action.h
@@ -6,8 +6,8 @@ public:
 	AOLIB_OBJECT IDENTITY GetIdentity() const;
 	AOLIB_OBJECT DWORD GetLockedSkillId1() const;
 	AOLIB_OBJECT DWORD GetLockedSkillId2() const;
-	//AOLIB_OBJECT DWORD GetLockoutTimeRemaining() const;
-	AOLIB_OBJECT float GetLockoutTimeRemaining() const;
+	AOLIB_OBJECT DWORD GetLockoutTimeRemaining() const;
+	//AOLIB_OBJECT float GetLockoutTimeRemaining() const;
 	AOLIB_OBJECT PCSTR GetName() const;
 	AOLIB_OBJECT SPECIALACTION GetSpecialActionData() const;
 	AOLIB_OBJECT IDENTITY GetWeaponIdentity() const;
